Reorganise tests from PatientExtensionTest to ResourceExtensionTest
- Add testExtractFamilyTag to PatientExtension
